requirements.txt
asgiref==3.2.10
chardet==3.0.4
Django==3.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-render-partial==0.4
django-taggit==1.3.0
jsmin==2.2.2
ldif3==3.1.1
Pillow==7.2.0
python-card-me==0.9.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
surlex==0.2.0
unicodecsv==0.14.1
